universe framework - cassandra deployment fixes
