+++ universe/src/main/java/org/corfudb/universe/api/deployment/vm/VmParams.java
@@ -33,6 +33,7 @@ public class VmParams<P extends NodeParams> implements DeploymentParams<P> {
 
     @Builder
     @Getter
+    @ToString
     public static class VsphereParams {
         /**
          * Default https://10.173.65.98/sdk
