+++ universe/src/main/java/org/corfudb/universe/infrastructure/docker/universe/group/cluster/DockerCassandraCluster.java
@@ -39,6 +39,7 @@ public class DockerCassandraCluster extends AbstractCluster<
             GenericGroupParams<CassandraServerParams, DockerContainerParams<CassandraServerParams>> cassandraParams) {
         super(cassandraParams, universeParams);
         this.docker = docker;
+        init();
     }
 
     @Override
