+++ universe/src/main/java/org/corfudb/universe/infrastructure/docker/universe/node/server/cassandra/DockerCassandraServer.java
@@ -6,6 +6,7 @@ import lombok.Builder.Default;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
+import org.corfudb.universe.api.common.IpAddress;
 import org.corfudb.universe.api.deployment.docker.DockerContainerParams;
 import org.corfudb.universe.api.universe.group.GroupParams.GenericGroupParams;
 import org.corfudb.universe.api.universe.node.Node;
@@ -87,6 +88,11 @@ public class DockerCassandraServer implements Node<CassandraServerParams, Docker
         });
 
         return this;
+    }
+
+    @Override
+    public IpAddress getNetworkInterface() {
+        return IpAddress.builder().ip(params.getName()).build();
     }
 
 }
